requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
sqlalchemy==2.0.27
python-multipart==0.0.9
pandas==2.2.0
duckduckgo-search==4.1.1
requests==2.31.0
requests-html==0.10.0
beautifulsoup4==4.12.3
feedparser==6.0.11
newspaper3k==0.2.8
python-dateutil==2.8.2
aiofiles==23.2.1
python-jose==3.3.0
passlib==1.7.4
undetected-chromedriver==3.5.5
motor==3.3.2
pymongo==4.6.1
webdriver-manager==4.0.1
urllib3<2.0.0
websockets<11.0,>=10.0
pydantic==2.6.1 